requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
curl_cffi==0.13.0
Django==5.2.6
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
frozendict==2.4.6
h11==0.16.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
multitasking==0.0.12
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
peewee==3.18.2
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==0.2.65
